Aggiunta cuore, modifiche alle funzioni, separazione funzioni in più cartelle
